refactor: button response types
