+++ packages/ai-chat/src/chat/components/ButtonItemCustomEventComponent.tsx
@@ -8,18 +8,15 @@
  */
 
 import TouchInteraction16 from "@carbon/icons/es/touch--interaction/16.js";
-import { carbonIconToReact } from "../../../utils/carbonIcon";
+import { carbonIconToReact } from "../utils/carbonIcon";
 import React, { useCallback } from "react";
-import { useSelector } from "../../../hooks/useSelector";
+import { useSelector } from "../hooks/useSelector";
 
-import { useServiceManager } from "../../../hooks/useServiceManager";
-import { selectInputIsReadonly } from "../../../store/selectors";
-import { LocalMessageItem } from "../../../../types/messaging/LocalMessageItem";
-import {
-  ButtonItem,
-  MessageResponse,
-} from "../../../../types/messaging/Messages";
-import { BusEventType } from "../../../../types/events/eventBusTypes";
+import { useServiceManager } from "../hooks/useServiceManager";
+import { selectInputIsReadonly } from "../store/selectors";
+import { LocalMessageItem } from "../../types/messaging/LocalMessageItem";
+import { ButtonItem, MessageResponse } from "../../types/messaging/Messages";
+import { BusEventType } from "../../types/events/eventBusTypes";
 import { BaseButtonItemComponent } from "./BaseButtonItemComponent";
 
 /**
