+++ packages/ai-chat/src/chat/components/ButtonItemPostBackComponent.tsx
@@ -8,20 +8,20 @@
  */
 
 import Send16 from "@carbon/icons/es/send/16.js";
-import { carbonIconToReact } from "../../../utils/carbonIcon";
+import { carbonIconToReact } from "../utils/carbonIcon";
 import React, { useCallback } from "react";
-import { useSelector } from "../../../hooks/useSelector";
+import { useSelector } from "../hooks/useSelector";
 
-import { useServiceManager } from "../../../hooks/useServiceManager";
-import { selectInputIsReadonly } from "../../../store/selectors";
-import { HasRequestFocus } from "../../../../types/utilities/HasRequestFocus";
-import { LocalMessageItem } from "../../../../types/messaging/LocalMessageItem";
-import { WA_CONSOLE_PREFIX } from "../../../utils/constants";
-import { createMessageRequestForButtonItemOption } from "../../../utils/messageUtils";
-import { consoleError } from "../../../utils/miscUtils";
-import actions from "../../../store/actions";
-import { MessageSendSource } from "../../../../types/events/eventBusTypes";
-import { ButtonItem } from "../../../../types/messaging/Messages";
+import { useServiceManager } from "../hooks/useServiceManager";
+import { selectInputIsReadonly } from "../store/selectors";
+import { HasRequestFocus } from "../../types/utilities/HasRequestFocus";
+import { LocalMessageItem } from "../../types/messaging/LocalMessageItem";
+import { WA_CONSOLE_PREFIX } from "../utils/constants";
+import { createMessageRequestForButtonItemOption } from "../utils/messageUtils";
+import { consoleError } from "../utils/miscUtils";
+import actions from "../store/actions";
+import { MessageSendSource } from "../../types/events/eventBusTypes";
+import { ButtonItem } from "../../types/messaging/Messages";
 import { BaseButtonItemComponent } from "./BaseButtonItemComponent";
 
 interface ButtonItemPostBackComponentProps extends HasRequestFocus {
